feature(youtube): Add youtube video content got from YT API
Youtube videos are part of my content, so I decided to put them in the
portfolio.

In the future, I can put more attention on API_KEY handling, but, since
this app goes completely to front, an special endpoint should be created
on Servlet to hide the keys. For now, It has been restricted via GCloud,
but there's nothing else to do at this point.